Updated for use with Python 3.12 Updated requirements to work with Python 3.12 Updated for use with Exid dataset version 2.1 Changed folder structure for HighD and ExiD to prevent having to copy all images before use

+++ visualisation/visualisationmaster.py
@@ -44,7 +44,7 @@ class VisualisationMaster(QtCore.QObject):
         self.path_to_video_file = ''
 
         self.main_timer = QtCore.QTimer()
-        self.main_timer.setInterval(self.dt.total_seconds() * 1e3)
+        self.main_timer.setInterval(int(self.dt.total_seconds() * 1e3))
         self.main_timer.setTimerType(QtCore.Qt.PreciseTimer)
         self.main_timer.setSingleShot(False)
         self.main_timer.timeout.connect(self._do_time_step_wrapper)
